Separate [basic ui components] ui/basics
