+++ pages/test/index.tsx
@@ -1,8 +1,5 @@
-import { Helper } from '@ui/Helper/Helper';
-import { InfoButton } from '@ui/InfoButton/InfoButton';
 import { DefaultLayout } from '@layouts/DefaultLayout';
-import { Button } from '@mantine/core';
-import { ReactElement, useCallback } from 'react';
+import { ReactElement } from 'react';
 
 function TestPage() {
   return <></>;
