+++ pages/user/list.tsx
@@ -15,7 +15,7 @@ import { useRequest } from '@hooks/useRequest';
 import { ILocale } from '@custom-types/ui/ILocale';
 import Fuse from 'fuse.js';
 import { hasSubarray } from '@utils/hasSubarray';
-import MultiSelect from '@ui/Select/MultiSelect';
+import { MultiSelect } from '@ui/basics';
 import { customTableSort } from '@utils/customTableSort';
 import { IUser, IUserListBundle } from '@custom-types/data/IUser';
 import { IGroup } from '@custom-types/data/IGroup';
