+++ src/components/AssignmentSchema/Form/TaskAdding/TaskAdding.tsx
@@ -3,7 +3,7 @@ import { FC, memo } from 'react';
 import TaskSelector from '../TaskSelector/TaskSelector';
 import styles from './taskAdding.module.css';
 import stepperStyles from '@styles/ui/stepper.module.css';
-import InputWrapper from '@ui/InputWrapper/InputWrapper';
+import { InputWrapper } from '@ui/basics';
 
 const TaskAdding: FC<{ form: any; initialTasks: Item[] }> = ({
   form,
