+++ src/components/ui/basics/SegmentedControl/SegmentedControl.tsx
