+++ src/components/ui/basics/SegmentedControl/segmentedControl.module.css
