+++ src/components/ui/basics/index.tsx
@@ -1,0 +1,33 @@
+import Button from './Button/Button';
+import CustomEditor from './CustomEditor/CustomEditor';
+import Dropzone from './Dropzone/Dropzone';
+import Helper from './Helper/Helper';
+import Icon from './Icon/Icon';
+import InputWrapper from './InputWrapper/InputWrapper';
+import NumberInput from './NumberInput/NumberInput';
+import MultiSelect from './Select/MultiSelect';
+import PasswordInput from './PasswordInput/PasswordInput';
+import Radio from './Radio/Radio';
+import SegmentedControl from './SegmentedControl/SegmentedControl';
+import Select from './Select/Select';
+import Switch from './Switch/Switch';
+import TextArea from './TextArea/TextArea';
+import TextInput from './TextInput/TextInput';
+
+export {
+  Button,
+  CustomEditor,
+  Dropzone,
+  Helper,
+  Icon,
+  InputWrapper,
+  MultiSelect,
+  NumberInput,
+  PasswordInput,
+  Radio,
+  SegmentedControl,
+  Select,
+  Switch,
+  TextArea,
+  TextInput,
+};
